app/requirements.txt
customtkinter>=5.2.2
pyserial>=3.5
matplotlib>=3.8.2
pandas>=2.2.0
numpy>=1.26.3
requests==2.26.0
Pillow>=10.2.0
openpyxl>=3.1.2
PyQt6>=6.6.1
flask==2.0.1
python-dotenv==0.19.0
pyinstaller==6.12.0 